refactor legacy
