+++ cocos/rendering/index.ts
@@ -34,18 +34,6 @@ export { RenderFlow } from './render-flow';
 export { RenderStage } from './render-stage';
 export { PipelineSceneData } from './pipeline-scene-data';
 
-export { ForwardPipeline, createDefaultPipeline } from './forward/forward-pipeline';
-export { ForwardFlow } from './forward/forward-flow';
-export { ForwardStage } from './forward/forward-stage';
-export { DeferredPipeline } from './deferred/deferred-pipeline';
-export { MainFlow } from './deferred/main-flow';
-export { GbufferStage } from './deferred/gbuffer-stage';
-export { LightingStage } from './deferred/lighting-stage';
-export { BloomStage } from './deferred/bloom-stage';
-export { PostProcessStage } from './deferred/postprocess-stage';
-export { ShadowFlow } from './shadow/shadow-flow';
-export { ShadowStage } from './shadow/shadow-stage';
-
 export { InstancedBuffer } from './instanced-buffer';
 export { PipelineStateManager } from './pipeline-state-manager';
